Move JSON deserialization and models to NewRelic.Core (#2378)
* Move Json deserialization from wrapper to core

* Also needed to move the models to core as well

* Add and tweak comments/readme
